Add `/metrics/<aggregate>` route and implement for count
Added new route for performing aggregations on a filterable
list of metrics. In this pass, it allows for normal filtering,
but only aggragates with counts.

A request to this enpoint is of the form:

GET /metrics/count?start_datetime=2020-01-01T:00:00:00
&end_datetime=2020-01-02T00:00:00&bucket_count=4

On a successful request, a list of buckets is returned, of the
format:

```
{
   "buckets": [
	{ "result": 6, "bucket": 0 },
        { "result": 2, "bucket": 5 ),
        ...
    ]
}
```

for all non-empty buckets.

NW