Add aggregation functionality for min, max, avg, and sum
This allows for queries to the `/metrics/<aggregation>` to specify
any of the above mentioned values (min, avg, max, or sum) and for that
operation to be performed on the resulting filtered data.

So, requests of the form:

GET
`/metrics/<aggregation>?bucket_count=x&start_datetime=xxx&end_datetime=xxx
&q=xxx&metric_name=xxx`

will return

```
{
	"data": {
		"buckets": [
			{"value": x, "bucket": "2019-01-01T00:00:00"},
			...
		]
	}
}
```

where `metric_name` specifies the path to the value in the metric data
to perform the aggregate on, and `value` is the resulting aggregate for
the given bucket.

NW